pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "snokpok", email = "vinceny.fe@gmail.com" }
]
dependencies = [
    "web3>=7.3.0",
    "flask>=3.0.3",
    "pymongo>=4.10.1",
    "python-dotenv>=1.0.1",
    "flask-cors>=5.0.0",
    "langchain-openai>=0.2.2",
    "py-solc-x>=2.0.3",
    "langgraph>=0.2.35",
    "langchain-community>=0.3.2",
    "langchain-experimental>=0.3.2",
    "langchain-anthropic>=0.2.3",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/api"]
